Se borra clase de prueba modal y se define componente final
