+++ front/src/App.js
@@ -2,11 +2,12 @@ import React from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 import PruebasTable from './components/pruebas/PruebasTable';
 import PruebasForm from './components/pruebas/PruebasForm';
-import PruebasModal from './components/pruebas/PruebaModal';
+import PruebasModal from './components/common/Modal';
 import Login from './components/login/Login';
 import Main from './components/Main';
 import Test from './components/common/Modal';
 import Vendedores from './components/Vendedores';
+import { Modal } from 'reactstrap';
 
 function App() {
   return (  
@@ -15,8 +16,7 @@ function App() {
         <Route exact path="/login" component={Login}></Route>
         <Route exact path="/tablas" component={PruebasTable}></Route>
         <Route exact path="/formulario" component={PruebasForm}></Route>
-        <Route exact path="/main" component={Main}></Route>
-        <Route exact path="/" component={Vendedores}></Route>
+        <Route exact path="/" component={Main}></Route>
       </Switch>
     </BrowserRouter>
   );
