+++ front/src/components/common/Modal.js
@@ -1,41 +1,53 @@
-import React, { useState } from 'react';
-import '../../styles/Modal.css';
-import Form from './Form';
-import Button from './Button';
+import React from 'react';
+import formulario from '../../resources/json/prueba.json';
+import  Form  from './Form';
+import {
+  Button,
+  Modal,
+  ModalHeader,
+  ModalBody,
+  FormGroup,
+  ModalFooter,
+} from "reactstrap";
 
-const Modal = ({title, formulario, visible}) => {
-    const [show, setShow] = useState(false);
-     
-    const handleModalClose = (e) => {
-      const currentClass = e.target.className;
-      
-      if (currentClass === 'modal-card') {
-        return;
-      }
-      
-      setShow(false);
-    };
-    
-    const handleModalOpen = () => {
-      setShow(true);
-    };
-    
-    return (
-        <div
-          hidden={!show}  
-        >
-            <div className="modal-card">
-                <h3>{title}</h3>
-                <hr class="line-title"/>
-                <Form fields={formulario} />
-                <div className="container-buttons">
-                  <Button label={"Cancelar"} name={"button"} event={handleModalClose}></Button>
-                  <Button label={"Aceptar"} name={"button"}></Button>
-                </div>
-            </div>
-         
-        </div>
-    );
-  };
-  
-  export default Modal;
+const PruebasTable = ({isOpen, handleOpen}) => {
+ 
+  // const [modalActualizar, setModalActualizar] = React.useState(true);
+
+  // const cerrarModalActualizar = () => {
+  //   setModalActualizar(false);
+  // };
+
+  // const mostrarModalActualizar = () => {
+  //   setModalActualizar(true);
+  // };
+
+  return (
+    <Modal isOpen={isOpen}>
+        <ModalHeader>
+          <div><h3>Actualizar Usuario </h3></div>
+        </ModalHeader>
+
+        <ModalBody>
+             <FormGroup>
+               <Form fields={formulario}></Form>
+             </FormGroup>
+        </ModalBody>
+
+        <ModalFooter>
+          <Button
+            color="primary"
+          >
+            Actualizar
+          </Button>
+          <Button
+            className="btn btn-danger"
+            onClick={handleOpen}
+          >
+            Cancelar
+          </Button>
+        </ModalFooter>
+      </Modal>
+  );
+  }
+export default PruebasTable;
